requirements.txt
django
djangorestframework
celery
requests
celery
pytest
coverage